starting figuring out how to not remove wrappers

+++ snakemake/executors.py
@@ -566,7 +566,7 @@ class ClusterExecutor(RealExecutor):
         with self.lock:
             self.wait = False
         self.wait_thread.join()
-        if not self.workflow.immediate_submit:
+        if not self.workflow.immediate_submit or self.workflow.cleanup_wrappers:
             # Only delete tmpdir (containing jobscripts) if not using
             # immediate_submit. With immediate_submit, jobs can be scheduled
             # after this method is completed. Hence we have to keep the
@@ -897,12 +897,20 @@ class GenericClusterExecutor(ClusterExecutor):
 
             def job_status(job):
                 if os.path.exists(active_job.jobfinished):
-                    os.remove(active_job.jobfinished)
-                    os.remove(active_job.jobscript)
+                    if self.workflow.cleanup_wrappers:
+                        logger.debug("--skip-wrapper-cleanup not set, cleaning.")
+                        print(active_job.job_jobfinished)
+                        print(active_job.job_jobscript)
+                        os.remove(active_job.jobfinished)
+                        os.remove(active_job.jobscript)
                     return success
                 if os.path.exists(active_job.jobfailed):
-                    os.remove(active_job.jobfailed)
-                    os.remove(active_job.jobscript)
+                    if self.workflow.cleanup_wrappers:
+                        logger.debug("--skip-wrapper-cleanup not set, cleaning.")
+                        print(active_job.job_jobfailed)
+                        print(active_job.job_jobscript)
+                        os.remove(active_job.jobfailed)
+                        os.remove(active_job.jobscript)
                     return failed
                 return running
 
@@ -1035,11 +1043,15 @@ class SynchronousClusterExecutor(ClusterExecutor):
                         still_running.append(active_job)
                     elif exitcode == 0:
                         # job finished successfully
-                        os.remove(active_job.jobscript)
+                        if self.workflow.cleanup_wrappers:
+                            logger.debug("--skip-wrapper-cleanup not set, cleaning.")
+                            os.remove(active_job.jobscript)
                         active_job.callback(active_job.job)
                     else:
                         # job failed
-                        os.remove(active_job.jobscript)
+                        if self.workflow.cleanup_wrappers:
+                            logger.debug("--skip-wrapper-cleanup not set, cleaning.")
+                            os.remove(active_job.jobscript)
                         self.print_job_error(active_job.job)
                         self.print_cluster_job_error(
                             active_job, self.dag.jobid(active_job.job)
@@ -1193,11 +1205,15 @@ class DRMAAExecutor(ClusterExecutor):
                             WorkflowError("DRMAA Error: {}".format(e)),
                             self.workflow.linemaps,
                         )
-                        os.remove(active_job.jobscript)
+                        if self.workflow.cleanup_wrappers:
+                            logger.debug("--skip-wrapper-cleanup not set, cleaning.")
+                            os.remove(active_job.jobscript)
                         active_job.error_callback(active_job.job)
                         continue
                     # job exited
-                    os.remove(active_job.jobscript)
+                    if self.workflow.cleanup_wrappers:
+                        logger.debug("--skip-wrapper-cleanup not set, cleaning.")
+                        os.remove(active_job.jobscript)
                     if (
                         not retval.wasAborted
                         and retval.hasExited
